refactor(server): extract createApp() factory from index.ts
Splits server entry point so tests can create an Express instance
without starting the server or attaching WebSocket.

- app.ts exports createApp() — pure Express setup
- index.ts calls createApp(), attaches WebSocket, runs migrations, starts server
